+++ src/definitions/README.md
@@ -1,0 +1,5 @@
+# Discord Definitions
+
+The contents of these files comes straight from the API docs - they should be a direct match.
+
+TODO: Describe more about this
