+++ src/definitions/api.ts
@@ -1,0 +1,8 @@
+/**
+ * These type definitions come from the official Discord API docs. They should
+ * be defined with references back to the documentation section.
+ */
+
+// https://discord.com/developers/docs/reference#api-reference-base-url
+// https://discord.com/developers/docs/reference#api-versioning-api-versions
+export const API_URL = "https://discord.com/api/v9";
