+++ src/definitions/guilds.ts
@@ -1,0 +1,7 @@
+/**
+ * These type definitions come from the official Discord API docs. They should
+ * be defined with references back to the documentation section.
+ */
+
+// FIXME: Implement this
+export type Guild = Record<string, unknown>;
