+++ src/definitions/index.ts
@@ -1,0 +1,15 @@
+/**
+ * These type definitions come from the official Discord API docs. They should
+ * be defined with references back to the documentation section.
+ */
+
+export * from "./api";
+export * from "./application-commands";
+export * from "./channels";
+export * from "./components";
+export * from "./emoji";
+export * from "./guild-members";
+export * from "./guilds";
+export * from "./roles";
+export * from "./snowflakes";
+export * from "./users";
