+++ src/interaction.ts
@@ -2,9 +2,9 @@ import type { FastifyReply, FastifyRequest } from "fastify";
 import {
 	ApplicationCommandInteractionDataOption,
 	Interaction as IInteraction,
-	InteractionCallbackResponse,
+	InteractionResponse,
 	InteractionType,
-} from "./api/api";
+} from "./definitions";
 import { PermissionFlags } from "./data/messages";
 
 export default class Interaction {
@@ -36,7 +36,7 @@ export default class Interaction {
 		message: string;
 		ephemeral: boolean;
 	}) {
-		const payload: InteractionCallbackResponse = {
+		const payload: InteractionResponse = {
 			type: 4,
 			data: { content: message },
 		};
