--- src/records/channel.ts
@@ -1,1 +1,0 @@
-export type Channel = Record<string, unknown>;
