--- src/records/member.ts
@@ -1,1 +1,0 @@
-export type GuildMember = Record<string, unknown>;
