--- src/records/role.ts
@@ -1,1 +1,0 @@
-export type Role = Record<string, unknown>;
