--- src/records/user.ts
@@ -1,1 +1,0 @@
-export type User = Record<string, unknown>;
